Tidy up new 'none_fill_value' option to df builder.
It's different from sparse_fill_value so it should have a different kwarg. sparse_fill_value is used to convert a dataframe to a more space efficient sparse dataframe.

Add none_fill_value to all types, not just when the initial value is None, and assume that if the first value of a node is None it's a scalar (increase the isinstance check rather than add a new special case for None).