Simplify passing nodes to nodetype functions.
Use the convention that if an argument ends with "_node" then it expects a node to be passed to it.
The first argument is always the target node or value. If the first arugment ends with "_node" the node is passed, otherwise the node value.
Positional arguments can now be used when calling the node type (either as a decorator or as a node method).